Gallery class was added.
