+++ src/photo.js
@@ -4,26 +4,23 @@ var elementToClone = require('./elementToClone')();
 var common = require('./common');
 var utils = require('./utils');
 var renderPictures = require('./renderPictures');
+var Gallery = require('./gallery');
 
 /** @type {number} */
 var renderedPictureCount = 0;
 
+
 /**
  *
  * @param {Object} data
- * @param {Object} allData
  * @param {HTMLElement} container
  * @constructor
  */
-function Photo(data, container, allData) {
-  this.allData = allData;
+function Photo(data, container) {
   this.data = data;
   this.element = container;
   this.img = null;
   this.prevIndex = 0;
-  this.overlay = document.querySelector('.gallery-overlay');
-  this._onDocumentKeyDown = _onDocumentKeyDown.bind(this);
-  this._onPhotoClick = _onPhotoClick.bind(this);
 }
 
 var fn = Photo.prototype;
@@ -49,43 +46,10 @@ fn.getPhotoElement = function() {
   element.querySelector('.picture-likes').textContent = this.data.likes;
 
   this.element.appendChild(element);
-  this.img.addEventListener('click', this.showGallery.bind(this));
+  var gallery = new Gallery();
+  this.img.addEventListener('click', gallery.showGallery.bind(gallery));
   common.renderedPictures.push(this.data);
   return element;
-};
-
-fn.showGallery = function(e) {
-  var target = e.target;
-
-  e.preventDefault();
-  this.prevIndex = getIndex(document.querySelector('.pictures'), target.closest('.picture'));
-  this.overlay.classList.remove('invisible');
-  this.overlay.addEventListener('click', this._onPhotoClick);
-  document.addEventListener('keydown', this._onDocumentKeyDown);
-  this.showPicture(this.prevIndex);
-};
-
-fn.removeGallery = function() {
-  this.overlay.classList.add('invisible');
-  this.overlay.removeEventListener('click', this._onPhotoClick);
-  this.overlay.removeEventListener('click', this.removeGallery);
-  document.removeEventListener('keydown', this._onDocumentKeyDown);
-};
-
-fn.showPicture = function(index) {
-  var picture = this.overlay.querySelector('.gallery-overlay-image'),
-    commentElement = this.overlay.querySelector('.comments-count'),
-    likesElement = this.overlay.querySelector('.likes-count');
-
-  var data = common.renderedPictures[index] || this.data;
-  picture.src = data.url;
-
-  this.setCount(commentElement, data.comments);
-  this.setCount(likesElement, data.likes);
-};
-
-fn.setCount = function(element, count) {
-  element.textContent = count;
 };
 
 function onLoadEndCallback() {
@@ -109,27 +73,5 @@ function errorCallback(element) {
   onLoadEndCallback();
 }
 
-function _onPhotoClick(evt) {
-  if (!evt.target.closest('.gallery-overlay-preview')) {
-    this.removeGallery();
-  }
-
-  if (evt.target.classList.contains('gallery-overlay-image')) {
-    this.showPicture(++this.prevIndex);
-  }
-}
-
-function _onDocumentKeyDown(evt) {
-  if (evt.keyCode === utils.keyCode.ESC) {
-    this.removeGallery();
-  }
-}
-
-function getIndex(node, el) {
-  var nodeList = Array.prototype.slice.call(node.children),
-    index = nodeList.indexOf(el);
-
-  return index;
-}
 
 module.exports = Photo;
